refactor: organize form types in types/form.d.ts
